Move tail to jsonl, the only place it is used.
